[examples] Extract common Linux platform app start up code
