+++ examples/all-clusters-app/linux/main.cpp
@@ -32,7 +32,7 @@
 #include <support/CHIPMem.h>
 #include <support/RandUtils.h>
 
-#include "Server.h"
+#include "AppMain.h"
 
 #include <cassert>
 #include <iostream>
@@ -54,28 +54,7 @@ bool emberAfBasicClusterMfgSpecificPingCallback(chip::app::Command * commandObj)
 
 int main(int argc, char * argv[])
 {
-    CHIP_ERROR err = CHIP_NO_ERROR;
-
-    err = chip::Platform::MemoryInit();
-    SuccessOrExit(err);
-
-    err = chip::DeviceLayer::PlatformMgr().InitChipStack();
-    SuccessOrExit(err);
-
-    // Init ZCL Data Model and CHIP App Server
-    InitServer();
-
-    // Init Mdns Server
-    app::Mdns::StartServer();
-
-    chip::DeviceLayer::PlatformMgr().RunEventLoop();
-
-exit:
-    if (err != CHIP_NO_ERROR)
-    {
-        std::cerr << "Failed to run All Clusters App: " << ErrorStr(err) << std::endl;
-        // End the program with non zero error code to indicate a error.
-        return 1;
-    }
+    VerifyOrDie(ChipLinuxAppInit(argc, argv) == 0);
+    ChipLinuxAppMainLoop();
     return 0;
 }
