+++ examples/platform/linux/AppMain.h
@@ -1,0 +1,4 @@
+#pragma once
+
+int ChipLinuxAppInit(int argc, char ** argv);
+void ChipLinuxAppMainLoop();
